LRSD-11523 combine restcontrollers by path
