+++ workspaces/liferay-customer-workspace/client-extensions/liferay-customer-etc-spring-boot/src/main/java/com/liferay/customer/AccountsRestController.java
@@ -9,9 +9,14 @@ import com.liferay.client.extension.util.spring.boot3.BaseRestController;
 import com.liferay.client.extension.util.spring.boot3.client.LiferayOAuth2AccessTokenManager;
 import com.liferay.customer.constants.HeatTagConstants;
 import com.liferay.customer.constants.JiraIssueConstants;
+import com.liferay.customer.model.AccountUsage;
 import com.liferay.customer.model.JiraSupportIssue;
 import com.liferay.customer.permission.BusinessEventPermission;
+import com.liferay.customer.service.GoogleCloudFunctionService;
 import com.liferay.customer.service.JiraService;
+import com.liferay.customer.service.KoroneikiService;
+import com.liferay.headless.admin.user.client.resource.v1_0.AccountResource;
+import com.liferay.osb.koroneiki.phloem.rest.client.dto.v1_0.ProductPurchase;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
@@ -38,29 +43,76 @@ import org.json.JSONObject;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.oauth2.jwt.Jwt;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
 
 /**
  * @author Jenny Chen
+ * @author Amos Fong
  */
+@RequestMapping("/accounts")
 @RestController
-public class AccountsSyncBusinessEventsRestController
-	extends BaseRestController {
+public class AccountsRestController extends BaseRestController {
 
-	@RequestMapping(
-		method = RequestMethod.POST,
-		path = "/accounts/{externalReferenceCode}/sync-business-events"
-	)
+	@GetMapping("/{externalReferenceCode}/usage")
+	public ResponseEntity<String> get(
+			@AuthenticationPrincipal Jwt jwt,
+			@PathVariable("externalReferenceCode") String externalReferenceCode)
+		throws Exception {
+
+		try {
+			_checkPermission(jwt, externalReferenceCode);
+
+			List<ProductPurchase> productPurchases =
+				_koroneikiService.searchProductPurchases(
+					"accountKey eq '" + externalReferenceCode +
+						"' and state eq 'Active'",
+					1, 1000, StringPool.BLANK);
+
+			JSONObject jsonObject =
+				_googleCloudFunctionService.fetchCustomerAccountUsage(
+					externalReferenceCode);
+
+			AccountUsage accountUsage = new AccountUsage(
+				productPurchases, jsonObject);
+
+			JSONObject accountUsageJSONObject = accountUsage.toJSONObject();
+
+			return new ResponseEntity<>(
+				accountUsageJSONObject.toString(), HttpStatus.OK);
+		}
+		catch (Exception exception) {
+			_log.error(exception, exception);
+
+			return new ResponseEntity(
+				exception.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+	}
+
+	@GetMapping("/{externalReferenceCode}/tickets")
+	public ResponseEntity<String> getTickets(
+			@AuthenticationPrincipal Jwt jwt,
+			@PathVariable("externalReferenceCode") String externalReferenceCode,
+			@RequestParam(defaultValue = "", required = false) String[]
+				ticketIds)
+		throws Exception {
+
+		return _getJSMTickets(jwt, externalReferenceCode, ticketIds);
+	}
+
+	@PostMapping("/{externalReferenceCode}/sync-business-events")
 	public ResponseEntity<String> post(
 			@AuthenticationPrincipal Jwt jwt,
 			@PathVariable("externalReferenceCode") String externalReferenceCode,
@@ -134,6 +186,20 @@ public class AccountsSyncBusinessEventsRestController
 				page += 1;
 			}
 		}
+	}
+
+	private void _checkPermission(Jwt jwt, String externalReferenceCode)
+		throws Exception {
+
+		AccountResource accountResource = AccountResource.builder(
+		).header(
+			HttpHeaders.AUTHORIZATION, "Bearer " + jwt.getTokenValue()
+		).endpoint(
+			lxcDXPMainDomain, lxcDXPServerProtocol
+		).build();
+
+		accountResource.getAccountByExternalReferenceCode(
+			externalReferenceCode);
 	}
 
 	private String _getAuthorization() {
@@ -271,6 +337,69 @@ public class AccountsSyncBusinessEventsRestController
 		return StringPool.BLANK;
 	}
 
+	private ResponseEntity<String> _getJSMTickets(
+			Jwt jwt, String externalReferenceCode, String[] ticketIds)
+		throws Exception {
+
+		try {
+			_businessEventPermission.check(
+				jwt, externalReferenceCode, ActionKeys.VIEW);
+
+			StringBundler sb = new StringBundler(12);
+
+			sb.append("Organization in aqlFunction('\"External Key\" = \"");
+			sb.append(externalReferenceCode);
+			sb.append("\"') and (status not in ('");
+			sb.append(
+				StringUtil.merge(
+					JiraIssueConstants.STATUSES_SOLVED_AND_CLOSED, "','"));
+			sb.append("')) and ");
+			sb.append(
+				JiraIssueConstants.toJQLCustomField(
+					_jiraSupportHCFieldRequestType));
+			sb.append(" = '");
+			sb.append(JiraIssueConstants.TYPE_GENERAL_REQUEST);
+			sb.append("'");
+
+			if (ArrayUtil.isNotEmpty(ticketIds)) {
+				sb.append(" or key in ('");
+				sb.append(StringUtil.merge(ticketIds, "','"));
+				sb.append("')");
+			}
+
+			List<JiraSupportIssue> jiraSupportIssues = _jiraService.search(
+				sb.toString(),
+				new String[] {"key", "labels", "status", "summary"});
+
+			JSONArray jsonArray = new JSONArray();
+
+			for (JiraSupportIssue jiraSupportIssue : jiraSupportIssues) {
+				jsonArray.put(_toJSONObject(jiraSupportIssue));
+			}
+
+			return new ResponseEntity<>(jsonArray.toString(), HttpStatus.OK);
+		}
+		catch (Exception exception) {
+			_log.error(exception, exception);
+
+			return new ResponseEntity(
+				exception.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+	}
+
+	private JSONObject _toJSONObject(JiraSupportIssue jiraSupportIssue) {
+		return new JSONObject(
+		).put(
+			"link", jiraSupportIssue.getTicketURL()
+		).put(
+			"status", jiraSupportIssue.getStatus()
+		).put(
+			"subject", jiraSupportIssue.getSummary()
+		).put(
+			"ticketId", jiraSupportIssue.getKey()
+		);
+	}
+
 	private void _updateAccountHeatTags(
 			Long accountId, String externalReferenceCode)
 		throws Exception {
@@ -403,16 +532,22 @@ public class AccountsSyncBusinessEventsRestController
 	private static final String _JSM_AUTOMATION_HEAT_TAG_SUFFIX = "_be";
 
 	private static final Log _log = LogFactory.getLog(
-		AccountsSyncBusinessEventsRestController.class);
+		AccountsRestController.class);
 
 	@Autowired
 	private BusinessEventPermission _businessEventPermission;
+
+	@Autowired
+	private GoogleCloudFunctionService _googleCloudFunctionService;
 
 	@Autowired
 	private JiraService _jiraService;
 
 	@Value("${liferay.customer.jira.support.hc.field.request.type}")
 	private String _jiraSupportHCFieldRequestType;
+
+	@Autowired
+	private KoroneikiService _koroneikiService;
 
 	@Autowired
 	private LiferayOAuth2AccessTokenManager _liferayOAuth2AccessTokenManager;
